# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Appliance Service Center Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List all customers
      parameters:
        - name: search
          in: query
          schema: { type: string }
        - name: paymentStatus
          in: query
          schema: { type: string }
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create a customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get customer by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "404":
          description: Not found
    patch:
      operationId: updateCustomer
      tags: [customers]
      summary: Update a customer
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerUpdate"
      responses:
        "200":
          description: Updated customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    delete:
      operationId: deleteCustomer
      tags: [customers]
      summary: Delete a customer
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
  /customers/{id}/history:
    get:
      operationId: getCustomerHistory
      tags: [customers]
      summary: Get full customer history (jobs, payments, appliances)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Customer history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerHistory"
  /customers/{id}/whatsapp-form:
    get:
      operationId: getCustomerWhatsappForm
      tags: [customers]
      summary: Get WhatsApp form link/data for customer
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: WhatsApp form data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WhatsappForm"

  # ── Appliances ──────────────────────────────────────────────
  /appliances:
    get:
      operationId: listAppliances
      tags: [appliances]
      summary: List all appliances
      parameters:
        - name: customerId
          in: query
          schema: { type: integer }
        - name: type
          in: query
          schema: { type: string }
      responses:
        "200":
          description: List of appliances
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appliance"
    post:
      operationId: createAppliance
      tags: [appliances]
      summary: Add an appliance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplianceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appliance"
  /appliances/{id}:
    get:
      operationId: getAppliance
      tags: [appliances]
      summary: Get appliance by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Appliance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appliance"
    patch:
      operationId: updateAppliance
      tags: [appliances]
      summary: Update an appliance
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplianceUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appliance"
    delete:
      operationId: deleteAppliance
      tags: [appliances]
      summary: Delete an appliance
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ── Service Jobs ──────────────────────────────────────────────
...
  /highlights:
    get:
      operationId: listHighlights
      tags: [jobs]
      summary: List highlights for a job or customer
      parameters:
        - name: jobId
          in: query
          schema: { type: integer }
        - name: customerId
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: List of highlights
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Highlight"
    post:
      operationId: createHighlight
      tags: [jobs]
      summary: Create a highlight
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HighlightInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Highlight"
  /highlights/{id}:
    patch:
      operationId: updateHighlight
      tags: [jobs]
      summary: Update a highlight (tick, size, label)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HighlightUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Highlight"
    delete:
      operationId: deleteHighlight
      tags: [jobs]
      summary: Delete a highlight
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ── Reminders ──────────────────────────────────────────────
...
components:
  schemas:
...
    Customer:
      type: object
      required: [id, serialNumber, name, phone, createdAt]
      properties:
        id: { type: integer }
        serialNumber: { type: integer }
        name: { type: string }
        phone: { type: string }
        whatsappPhone: { type: ["string", "null"] }
        address: { type: ["string", "null"] }
        dpUrl: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        totalJobs: { type: integer }
        unpaidAmount: { type: number }
        lastJobDate: { type: ["string", "null"] }
        createdAt: { type: string }
    CustomerInput:
      type: object
      required: [name, phone]
      properties:
        name: { type: string, minLength: 1 }
        phone: { type: string, minLength: 1 }
        whatsappPhone: { type: string }
        address: { type: string }
        dpUrl: { type: string }
        notes: { type: string }
    CustomerUpdate:
      type: object
      properties:
        name: { type: string }
        phone: { type: string }
        whatsappPhone: { type: string }
        address: { type: string }
        dpUrl: { type: string }
        notes: { type: string }
    CustomerHistory:
      type: object
      required: [customer, jobs, appliances, totalPaid, totalDue]
      properties:
        customer:
          $ref: "#/components/schemas/Customer"
        jobs:
          type: array
          items:
            $ref: "#/components/schemas/ServiceJob"
        appliances:
          type: array
          items:
            $ref: "#/components/schemas/Appliance"
        totalPaid: { type: number }
        totalDue: { type: number }
    WhatsappForm:
      type: object
      required: [whatsappLink, messageTemplate]
      properties:
        whatsappLink: { type: string }
        messageTemplate: { type: string }
        customerName: { type: string }
        phone: { type: string }

    # ── Appliance ──────────────────────────────────────────────
    Appliance:
      type: object
      required: [id, customerId, type, createdAt]
      properties:
        id: { type: integer }
        customerId: { type: integer }
        customerName: { type: ["string", "null"] }
        type: { type: string }
        brand: { type: ["string", "null"] }
        model: { type: ["string", "null"] }
        serialNo: { type: ["string", "null"] }
        purchaseDate: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        createdAt: { type: string }
    ApplianceInput:
      type: object
      required: [customerId, type]
      properties:
        customerId: { type: integer }
        type: { type: string }
        brand: { type: string }
        model: { type: string }
        serialNo: { type: string }
        purchaseDate: { type: string }
        notes: { type: string }
    ApplianceUpdate:
      type: object
      properties:
        type: { type: string }
        brand: { type: string }
        model: { type: string }
        serialNo: { type: string }
        purchaseDate: { type: string }
        notes: { type: string }

    # ── ServiceJob ──────────────────────────────────────────────
    ServiceJob:
      type: object
      required: [id, jobNumber, customerId, status, paymentStatus, createdAt]
      properties:
        id: { type: integer }
        jobNumber: { type: string }
        customerId: { type: integer }
        customerName: { type: ["string", "null"] }
        customerPhone: { type: ["string", "null"] }
        applianceId: { type: ["integer", "null"] }
        applianceType: { type: ["string", "null"] }
        description: { type: ["string", "null"] }
        status:
          type: string
          enum: [pending, in_progress, completed, cancelled]
        paymentStatus:
          type: string
          enum: [unpaid, partial, paid]
        amount: { type: number }
        paidAmount: { type: number }
        technicianName: { type: ["string", "null"] }
        scheduledDate: { type: ["string", "null"] }
        completedDate: { type: ["string", "null"] }
        isHighlighted: { type: boolean }
        isNumberHighlighted: { type: boolean }
        highlights:
          type: array
          items:
            $ref: "#/components/schemas/Highlight"
        createdAt: { type: string }
        updatedAt: { type: string }
...
    Highlight:
      type: object
      required: [id, label, color, createdAt]
      properties:
        id: { type: integer }
        jobId: { type: ["integer", "null"] }
        customerId: { type: ["integer", "null"] }
        label: { type: string }
        color: { type: string }
        captionSize: { type: number }
        isNumbered: { type: boolean }
        isTicked: { type: boolean }
        zoomLevel: { type: number }
        createdAt: { type: string }
    HighlightInput:
      type: object
      required: [label, color]
      properties:
        jobId: { type: integer }
        customerId: { type: integer }
        label: { type: string }
        color: { type: string }
        captionSize: { type: number }
        isNumbered: { type: boolean }
        isTicked: { type: boolean }
        zoomLevel: { type: number }
    HighlightUpdate:
      type: object
      properties:
        label: { type: string }
        color: { type: string }
        captionSize: { type: number }
        isNumbered: { type: boolean }
        isTicked: { type: boolean }
        zoomLevel: { type: number }

    # ── Reminder ──────────────────────────────────────────────
...