# lib/api-zod/src/generated/types/applianceInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Appliance Service Center Management API
 * OpenAPI spec version: 0.1.0
 */

export interface ApplianceInput {
  customerId: number;
  type: string;
  brand?: string;
  model?: string;
  serialNo?: string;
  purchaseDate?: string;
  notes?: string;
}