# lib/api-zod/src/generated/types/customerHistory.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Appliance Service Center Management API
 * OpenAPI spec version: 0.1.0
 */
import type { Appliance } from './appliance';
import type { Customer } from './customer';
import type { ServiceJob } from './serviceJob';

export interface CustomerHistory {
  customer: Customer;
  jobs: ServiceJob[];
  appliances: Appliance[];
  totalPaid: number;
  totalDue: number;
}