# lib/api-zod/src/generated/types/highlightUpdate.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Appliance Service Center Management API
 * OpenAPI spec version: 0.1.0
 */

export interface HighlightUpdate {
  label?: string;
  color?: string;
  captionSize?: number;
  isNumbered?: boolean;
  isTicked?: boolean;
  zoomLevel?: number;
}